+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.1 (TBD)
 - Remove support for non-uikit themes
+- Fix call to geocoder function from UserLoc class.
 
 ## 1.2.0 (2018-11-03)
 - Add separate Google API key for Javascript