+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Remove support for non-uikit themes
   * Fix call to geocoder function from UserLoc class.
   * Accept width and height parameters in autotags.
+  * Add country field for markers; hopefully improve free-form address resolution.
 
 ## 1.2.0 (2018-11-03)
 - Add separate Google API key for Javascript