+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # glLocator - Change Log
 
-## 1.2.1 (TBD)
+## 1.2.1 (2020-07-09)
   * Remove support for non-uikit themes
   * Fix call to geocoder function from UserLoc class.
   * Accept width and height parameters in autotags.